pyproject.toml
[build-system]
build-backend = "mesonpy"
requires = ["meson-python", "setuptools_scm[toml]>=6.0"]

[project]
name = "python-gmp"
dynamic = ["version"]
description = "Safe bindings to the GNU GMP library"
keywords = ["gmp", "multiple-precision", "arbitrary-precision", "bignum"]
license = "MIT"
license-files = ["LICENSE"]
readme = "README.rst"
authors = [{name = "Sergey B Kirpichev", email = "skirpichev@gmail.com"}]
maintainers = [{name = "Sergey B Kirpichev", email = "skirpichev@gmail.com"}]
classifiers = ["Development Status :: 4 - Beta",
               "Intended Audience :: Developers",
               "Intended Audience :: Science/Research",
               "Natural Language :: English",
               "Operating System :: POSIX",
               "Programming Language :: C",
               "Programming Language :: Python :: 3",
               "Programming Language :: Python :: 3 :: Only",
               "Programming Language :: Python :: 3.9",
               "Programming Language :: Python :: 3.10",
               "Programming Language :: Python :: 3.11",
               "Programming Language :: Python :: 3.12",
               "Programming Language :: Python :: 3.13",
               "Programming Language :: Python :: 3.14",
               "Programming Language :: Python :: Free Threading :: 2 - Beta",
               "Programming Language :: Python :: Implementation :: CPython",
               "Programming Language :: Python :: Implementation :: PyPy",
               "Programming Language :: Python :: Implementation :: GraalPy",
               "Operating System :: Microsoft :: Windows",
               "Operating System :: POSIX",
               "Operating System :: Unix",
               "Operating System :: MacOS",
               "Topic :: Scientific/Engineering :: Mathematics",
               "Topic :: Software Development :: Libraries :: Python Modules"]
requires-python = ">= 3.9"

[project.urls]
Homepage = "https://github.com/diofant/python-gmp"
"Source Code" = "https://github.com/diofant/python-gmp"
"Bug Tracker" = "https://github.com/diofant/python-gmp/issues"
Documentation = "https://python-gmp.readthedocs.io/en/latest/"

[tool.setuptools_scm]

[project.optional-dependencies]
tests = ["pytest", "hypothesis", "mpmath>=1.4.0a6"]
docs = ["sphinx>=8.2"]
develop = ["python-gmp[tests,docs]", "pre-commit", "pyperf"]

[tool.pytest.ini_options]
addopts = "--durations=7"
xfail_strict = true

[tool.ruff]
line-length = 79

[tool.ruff.lint]
select = ["E", "F", "I", "PT", "W", "Q000"]
ignore = ["PT011"]

[tool.cibuildwheel]
enable = "pypy cpython-prerelease cpython-freethreading graalpy"
skip = """*-win32 *-manylinux_i686 *-musllinux_* \
          *-manylinux_armv7l pp310* pp*aarch64* gp3*win*"""
before-all = "bash scripts/cibw_before_all.sh"
test-extras = "tests"
test-command = "pytest {package}/tests"
repair-wheel-command = """bash scripts/cibw_repair_wheel_command.sh \
                          {dest_dir} {wheel}"""

[tool.cibuildwheel.windows]
before-all = "msys2 -c scripts/cibw_before_all.sh"
before-build = "pip install wheel delvewheel"
repair-wheel-command = """delvewheel repair -w {dest_dir} {wheel} \
                          --add-path .local/bin --no-mangle-all \
                          --include-imports"""

[tool.cibuildwheel.environment]
LD_LIBRARY_PATH = "$(pwd)/.local/lib:$LD_LIBRARY_PATH"
PKG_CONFIG_PATH = "$(pwd)/.local/lib/pkgconfig"
PYTEST_ADDOPTS = "--verbose"
CFLAGS = "-Wall -Wpedantic -Werror -std=c17"

[tool.cibuildwheel.windows.environment]
PYTEST_ADDOPTS = "--verbose"
